Refactor types into respective directory
