+++ lib/base/index.js
@@ -2,13 +2,14 @@
 import {keybaseExec} from '../utils'
 import {withStore, actionInit} from '../store'
 import type {Store} from '../store/types'
-import type {InitOptions, DeviceUsernamePair, UsernameAndDevice} from '../types.js'
+import type {DeviceUsernamePair} from '../types'
+import type {InitOptions} from './types'
 
 /**
  * @ignore
  * _getCurrentUsernameAndDevicename returns { username, devicename } from keybase status -j
  */
-const _getCurrentUsernameAndDevicename = async (): Promise<UsernameAndDevice> => {
+const _getCurrentUsernameAndDevicename = async (): Promise<DeviceUsernamePair> => {
   const status = await keybaseExec(['status', '--json'], {json: true})
   if (status && status.Username && status.Device && status.Device.name) {
     return {username: status.Username, devicename: status.Device.name}
