+++ lib/base/types.js
@@ -1,0 +1,30 @@
+// @flow
+import type {DeviceUsernamePair} from '../types'
+
+/**
+ */
+export type InitOptions = {
+  username: string,
+  paperkey: string,
+  verbose?: boolean,
+}
+
+export type UsernameAndDevice = {
+  username: string,
+  devicename: string,
+}
+
+/*
+ * Base API functions
+ */
+
+export type Init = InitOptions => Promise<void>
+
+export type Deinit = () => Promise<void>
+
+export type MyInfo = () => ?DeviceUsernamePair
+
+/*
+ * Any Base API function
+ */
+export type AnyBaseApi = Init | Deinit | MyInfo
