+++ lib/chat/index.js
@@ -4,6 +4,7 @@ import {formatOptions, keybaseExec} from '../utils'
 import {CHAT_API_VERSION} from '../constants'
 import {withStore, withGuards} from '../store'
 import type {Store} from '../store/types'
+import type {ApiCommandArg} from '../types'
 import type {
   ChatChannel,
   ChatOptionsList,
@@ -11,9 +12,8 @@ import type {
   ChatReadMessage,
   ChatOptionsDelete,
   ChatOptionsRead,
-  ApiCommandArg,
   ChatConversation,
-} from '../types.js'
+} from './types'
 
 const _runApiCommand = async (arg: ApiCommandArg): Promise<any> => {
   // await guardInitialized()
